+++ src/tiffvisitor_int.cpp
@@ -456,7 +456,7 @@ namespace Exiv2 {
         // create vector of signedShorts from unsignedShorts in Exif.Canon.AFInfo
         std::vector<int16_t>  ints;
         std::vector<uint16_t> uint;
-        for (int i = 0; i < object->pValue()->count(); i++) {
+        for (long i = 0; i < object->pValue()->count(); i++) {
             ints.push_back(static_cast<int16_t>(object->pValue()->toLong(i)));
             uint.push_back(static_cast<uint16_t>(object->pValue()->toLong(i)));
         }
@@ -505,10 +505,10 @@ namespace Exiv2 {
                 auto v = Exiv2::Value::create(record.bSigned ? Exiv2::signedShort : Exiv2::unsignedShort);
                 std::ostringstream s;
                 if (record.bSigned) {
-                    for (int16_t k = 0; k < record.size; k++)
+                    for (uint16_t k = 0; k < record.size; k++)
                         s << " " << ints.at(nStart++);
                 } else {
-                    for (int16_t k = 0; k < record.size; k++)
+                    for (uint16_t k = 0; k < record.size; k++)
                         s << " " << uint.at(nStart++);
                 }
